py_server/requirements.txt
# Use Python 3.11 or 3.12 (3.13 not yet supported by this stack)
strawberry-graphql==0.226.2
uvicorn[standard]==0.30.1
starlette==0.37.2
